Created new namespace "Management" for storing static class management systems. Moved ProfileManager to new namespace. Moved GraphicsManager to new namespace. Added SettingsManager to Management namespace. Updating usings program wide.
