+++ SA-Mod-Manager/Management/GraphicsManager.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace SAModManager.Configuration
+namespace SAModManager.Management
 {
 	public static class GraphicsManager
 	{
