+++ SA-Mod-Manager/Management/SettingsManager.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using SAModManager.Configuration;
+
+namespace SAModManager.Management
+{
+	/// <summary>
+	/// Static Management class for the <see cref="ManagerSettings"/> used in the application.
+	/// </summary>
+    public static class SettingsManager
+    {
+
+    }
+}
