+++ SA-Mod-Manager/Profile/EditProfile.xaml.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Windows;
 using System.Windows.Input;
 using SAModManager.UI;
+using SAModManager.Management;
 
 namespace SAModManager.Profile
 {
